requirements.txt
# Web框架
fastapi==0.135.3
starlette==1.0.0
uvicorn==0.44.0
python-multipart==0.0.12

# 数据库
SQLAlchemy==2.0.49
pymysql==1.1.1

# 数据验证
pydantic==2.12.5
pydantic_settings==2.13.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 日志
loguru==0.7.3

# 数据可视化
matplotlib==3.10.8

# HTTP客户端
httpx==0.28.1

# AI/大模型
openai>=1.0.0
dashscope>=1.20.0

# 邮件
aiosmtplib>=3.0.0
email-validator>=2.2.0

# 模板
jinja2>=3.1.4
aiofiles>=24.1.0

# 配置
python-dotenv>=1.0.0

# 定时任务
apscheduler>=3.10.0
